airflow/dags/migrate_postgres_to_sql/requirements.txt
greenlet==1.1.2
numpy==1.23.1
pandas==1.4.3
protobuf==3.20.1
psycopg2-binary==2.9.3
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39